LengowCodingame/requirements.txt
requests>=2.26
flake8
djangorestframework
markdown
django-filter
regex